+++ __tests__/cases/server/server.js
@@ -108,6 +108,7 @@ describe('Server instance', () => {
 		expect(server.path).to.be.equal(path.resolve(PROJECT_PATH, `${staticDir}/${jsDir}`));
 		expect(new server.Layout()).to.be.an.instanceOf(Layout);
 		expect(new server.Session()).to.be.an.instanceOf(Session);
+		expect(server.Session).not.to.be.equal(null);
 	});
 
 	it('checks the set config fields of the server', () => {
@@ -197,6 +198,7 @@ describe('Server instance', () => {
 		expect(server.path).to.be.equal(path.resolve(PROJECT_PATH, `${staticDir}/${jsDir}`));
 		expect(new server.Layout()).to.be.an.instanceOf(Layout);
 		expect(new server.Session()).to.be.an.instanceOf(Session);
+		expect(server.Session).not.to.be.equal(null);
 	});
 
 	it('tries to set not Layout child as a layoutComponent', () => {